Add a loop close SSA transformation util.
A loop is in the loop close SSA form (LCSSA) if all out-of-loop uses
of value defined inside a loop go are phi instruction in loop exit blocks.
A loop exit block is a basic block that does not belong to the loop and
for which all its predecessors are in the loop.

The transformation follow the following steps:
 - Creates dedicate loop exit block if needed;
 - For each instruction in the loop that have a out-of-loop use:
   - Create a phi instruction for each exit block the instruction dominates;
   - Rewrite all out-of-loop uses in terms of those new phi instruction.